// src/app/models/Tesoreria/instrumentoPorCoberturar.ts
export class InstrumentoPorCoberturar{
    t463_id: number;
    t463_code_bbg: number;
    t463_currency_rec: string;
    t463_currency_del: string;
    t463_nominal_del: number;
    t463_nominal_rec: number;
    t463_id_subsidiary: number;
    t463_end_date: number;
    t463_id_counterpart: number;
    desc_counterpart_type: string;
    desc_subsidiary: string;
    subsidiary_country: string;
    desc_counterpart: string;
    t459_deal_code: string;
    desc_typeOperation: string;
    t463_registered_by: string;
    t463_registration_date: number;
    t463_record_type: string;
    t463_external_reference: string;
    t463_id_type_operation: number;
    t463_id_modality: string;
    t463_currency_liqu: string;
    t463_strike: number;
    t463_spot: number;
    t463_id_quotation: number;
    t463_start_date: number;
    t463_settlement_date: number;

    montoCubierto: number;
    flgInstrumentoSeleccionado: boolean;
    t463_type_ci: string;
    t463_trader_name: string;
    t463_counterparty_trader_name: string;
    t463_id_modality_supplementary: number;
    desc_modality_supplementary: string;

    usuario: string;
}